pyproject.toml
[project]
name = "lok-sabha-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling>=2.73.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "openai>=2.21.0",
    "pymupdf>=1.27.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.0.0",
    "qdrant-client[fastembed]>=1.16.2",
    "sse-starlette>=3.2.0",
    "typer>=0.21.2",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lok_sabha_rag"]
